refactor(desktop/preload): move InstallDependenciesResult type into preload/types
